Update ID copy
Previously, you had to hold `Ctrl` to see the ID and if you wanted to copy the ID after the tooltip was shown you had to leave the frame and enter it again while holding both `Ctrl` and `Alt`. Now, as long as the ID is shown in the tooltip you can just hold `Alt` to open popup and copy it so it's more intuitive.